+++ app/schemas/question.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 from datetime import datetime
 from enum import Enum
+from pydantic import field_validator
 from .base_config import BaseModelConfig
 from .pagination import Pagination
 from .solution import Solution
@@ -17,7 +18,7 @@ class Question(BaseModelConfig):
     test_cases: list[TestCase]
     notes: list[str]
     hints: list[str]
-    tags: list[str]
+    tags: list[Tag]
     solutions: list[Solution]
     created_at: datetime
     last_modified: datetime
@@ -33,7 +34,19 @@ class QuestionCreate(BaseModelConfig):
     test_cases: list[TestCase]
     notes: list[str]
     hints: list[str]
-    tags: list[str]
+    tags: list[Tag]
+
+    @field_validator("title")
+    def title_exists_and_is_long_enough(title: str) -> int:
+        if not (0 < len(title) <= 50):
+            raise ValueError("There must be a title and it should be less or equal than 50 characters.")
+        return title
+
+    @field_validator("prompt")
+    def prompt_exists(prompt: str) -> int:
+        if len(prompt) == 0:
+            raise ValueError("There must be a prompt for a question.")
+        return prompt
 
 
 class QuestionBasicInfo(BaseModelConfig):
@@ -42,7 +55,7 @@ class QuestionBasicInfo(BaseModelConfig):
     difficulty: Difficulty
     status: Status
     title: str
-    tags: list[str]
+    tags: list[Tag]
     created_at: datetime
     last_modified: datetime
 
@@ -50,6 +63,7 @@ class QuestionBasicInfo(BaseModelConfig):
 class QuestionAll(BaseModelConfig):
     paginated: bool
     data: Pagination | list[QuestionBasicInfo]
+
 
 class Source(Enum):
     LEETCODE = "leetcode"
@@ -66,3 +80,41 @@ class Difficulty(Enum):
     EASY = "easy"
     MEDIUM = "medium"
     HARD = "hard"
+
+
+from enum import Enum
+
+class Tag(Enum):
+    ARRAY = "array"
+    STRING = "string"
+    HASH_TABLE = "hash table"
+    DYNAMIC_PROGRAMMING = "dynamic programming"
+    MATH = "math"
+    SORTING = "sorting"
+    GREEDY = "greedy"
+    DFS = "dfs"
+    BINARY_SEARCH = "binary search"
+    MATRIX = "matrix"
+    BFS = "bfs"
+    TREE = "tree"
+    BIT_MANIPULATION = "bit manipulation"
+    TWO_POINTERS = "two pointers"
+    HEAP = "heap"
+    BINARY_TREE = "binary tree"
+    SIMULATION = "simulation"
+    STACK = "stack"
+    GRAPH = "graph"
+    SLIDING_WINDOW = "sliding window"
+    DESIGN = "design"
+    BACKTRACKING = "backtracking"
+    DISJOINT_SET = "disjoint set"
+    LINKED_LIST = "linked list"
+    TRIE = "trie"
+    QUEUE = "queue"
+    RECURSION = "recursion"
+    DIVIDE_AND_CONQUER = "divide and conquer"
+    BINARY_SEARCH_TREE = "binary search tree"
+    RANDOMIZED = "randomized"
+    ITERATOR = "iterator"
+    CONCURRENCY = "concurrency"
+    SHELL = "shell"