+++ app/schemas/token.py
@@ -1,11 +1,6 @@
-from datetime import datetime
 from .base_config import BaseModelConfig
 
 class Token(BaseModelConfig):
     iss: str
     aud: str
     exp: float
-
-
-class AuthenticateReq(BaseModelConfig):
-    secret: str